removed signout controller and placed in auth controller
