+++ controllers/auth.js
@@ -52,6 +52,18 @@ const githubAuth = (req, res, next) => {
   }
 };
 
+const signout = (req, res) => {
+  const cookies = Object.keys(req.cookies);
+  cookies.forEach((cookie) => {
+    res.clearCookie(cookie);
+  });
+  return res.json({
+    message: "Signout succesful",
+    deletedCookies: Object.keys(req.cookies),
+  });
+};
+
 module.exports = {
   githubAuth,
+  signout,
 };
